+++ message_ix_models/project/sparccle/cli.py
@@ -1,4 +1,4 @@
-"""Command-line tools specific to the SPARRCLE project."""
+"""Command-line tools specific to the SPARCCLE project."""
 
 from pathlib import Path
 
@@ -7,17 +7,17 @@ import click
 from message_ix_models.workflow import make_click_command
 
 
-@click.group("sparrcle")
+@click.group("sparccle")
 @click.pass_obj
 def cli(context):
-    """SPARRCLE project."""
+    """SPARCCLE project."""
 
 
 cli.add_command(
     make_click_command(
         f"{__package__}.workflow.generate",
-        name="SPARRCLE",
-        slug="sparrcle",
+        name="SPARCCLE",
+        slug="sparccle",
         params=[
             click.Option(
                 ["--config", "config_path"],